### 2017-04-06 #227 @whedon
[PRE REVIEW]: libqcpp: A C++14 sequence quality control library
**Submitting author:** @kdmurray91 (<a href="http://orcid.org/0000-0002-2466-1917">Kevin Murray</a>)
**Repository:** <a href="https://github.com/kdmurray91/libqcpp" target ="_blank">https://github.com/kdmurray91/libqcpp</a>
**Version:** 0.4.0
