### 2017-04-07 #227 @arfon
@luizirber please head over to #232 to carry out the review.
