### 2017-05-04 kdmurray91/libqcpp#3 @luizirber
JOSS review feedback
Hi!

While reviewing  https://github.com/openjournals/joss-reviews/issues/232 I missed some things:

- `Example usage` and `functionality`: an example (maybe using the test data you already have available?) helps on checking how the software works, also because the `trimit` help and [docs](http://qcpp.readthedocs.io/en/latest/applications.html#trimit) are a bit sparse.
- `Community guidelines`: you can add CONTRIBUTING.md to the repo, or expand the developer section of the docs

Nice API docs and library, BTW!

# Tasks

- [x] Contributing.md
- [x] Tutorial documentation
