### 2017-05-04 kdmurray91/libqcpp#3 @kdmurray91
@luizirber Thanks heaps for the review. I'm at a conference next week, but will do my best to have a small example & some tutorial-style docs prepared within a week. I'll put a contributing.md up too.
