### 2017-05-22 kdmurray91/libqcpp#3 @kdm9
Hi @luizirber, I think I have addressed your concerns. The online docs at qcpp.rtfd.io now contain a tutorial, and `CONTRIBUTING.md` now exists.
