### 2017-05-22 #232 @kdmurray91
Hi all, I've pushed changes that hopefully address all the review comments, as detailed in https://github.com/kdmurray91/libqcpp/issues/3.
