### 2017-05-26 #232 @kdmurray91
Hi @arfon

I've cut a new release, 0.4.1, that incorporates @luizirber's comments. The Zenodo doi for this is  http://doi.org/10.5281/zenodo.583325

Cheers,
Kevin
