### 2017-05-26 #232 @arfon
@luizirber - many thanks for the review here ✨

@kdmurray91 - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00232 ⚡️:rocket: 💥
